REFACTORING: Suggestions for topic assignment generalized.
